Optimized some queries

+++ src/main/java/com/example/magicmondays/DAO/IDeckRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IDeckRepository extends JpaRepository<Deck,Long> {
+public interface IDeckRepository extends JpaRepository<Deck,Integer> {
 }
