+++ src/main/java/com/example/magicmondays/DAO/IPlayerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPlayerRepository extends JpaRepository<Player,Long> {
+public interface IPlayerRepository extends JpaRepository<Player,Integer> {
 }
