+++ src/main/java/com/example/magicmondays/DAO/IRecordRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IRecordRepository extends JpaRepository<Record,Long> {
+public interface IRecordRepository extends JpaRepository<Record,Integer> {
 }
